+++ lib/handlers/toast_handler.dart
@@ -14,6 +14,7 @@ class ToastHandler extends ReportHandler {
   final Color textColor;
   final double textSize;
   final String? customMessage;
+  FToast? fToast;
 
   ToastHandler({
     this.gravity = ToastHandlerGravity.bottom,
@@ -23,8 +24,6 @@ class ToastHandler extends ReportHandler {
     this.textSize = 12,
     this.customMessage,
   });
-
-  FToast? fToast;
 
   @override
   Future<bool> handle(Report error, BuildContext? buildContext) async {
@@ -100,11 +99,13 @@ class ToastHandler extends ReportHandler {
 
   @override
   List<PlatformType> getSupportedPlatforms() => [
-        PlatformType.web,
-        PlatformType.android,
-        PlatformType.iOS,
-      ];
-
+    PlatformType.android,
+    PlatformType.iOS,
+    PlatformType.web,
+    PlatformType.linux,
+    PlatformType.macOS,
+    PlatformType.windows,
+  ];
   @override
   bool isContextRequired() {
     return true;
@@ -129,19 +130,22 @@ class FlutterToastPage extends StatefulWidget {
 }
 
 class _FlutterToastPageState extends State<FlutterToastPage> {
-  FToast fToast = FToast();
+  FToast _fToast = FToast();
+  bool _disposed = false;
 
   @override
   void initState() {
-    fToast.init(context);
+    _fToast.init(context);
     WidgetsBinding.instance!.addPostFrameCallback((_) {
-      showToast();
+      if (!_disposed && mounted) {
+        showToast();
+      }
     });
     super.initState();
   }
 
   void showToast() {
-    fToast.showToast(
+    _fToast.showToast(
         child: Container(
           color: widget.backgroundColor,
           child: Text(
@@ -157,7 +161,9 @@ class _FlutterToastPageState extends State<FlutterToastPage> {
     Future.delayed(
       Duration(milliseconds: widget.duration.inMilliseconds + 100),
       () {
-        Navigator.of(context).pop();
+        if (!_disposed && mounted) {
+          Navigator.of(context).pop();
+        }
       },
     );
   }
@@ -165,5 +171,11 @@ class _FlutterToastPageState extends State<FlutterToastPage> {
   @override
   Widget build(BuildContext context) {
     return const SizedBox();
+  }
+
+  @override
+  void dispose() {
+    _disposed = true;
+    super.dispose();
   }
 }