+++ lib/core/catcher.dart
@@ -528,7 +528,6 @@ class Catcher with ReportModeAction {
           "Error: '$error' has been filtered from Catcher logs. Report will be skipped.");
       return;
     }
-
     _cachedReports.add(report);
     ReportMode? reportMode =
         _getReportModeFromExplicitExceptionReportModeMap(error);
@@ -648,6 +647,12 @@ class Catcher with ReportModeAction {
 
   @override
   void onActionRejected(Report report) {
+    _currentConfig.handlers
+        .where((handler) => handler.shouldHandleWhenRejected())
+        .forEach((handler) {
+      _handleReport(report, handler);
+    });
+
     _cachedReports.remove(report);
   }