+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Updated chinese translations in LocalizationOptions.
 * Updated example app codebase.
 * Added filterFunction to CatcherOptions.
+* Fixed displaying toast in desktops. Added missing support for desktops in toast handler.
 
 ## 0.6.5
 * Changed package_info to package_info_plus to support all platforms.