+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Updated example app codebase.
 * Added filterFunction to CatcherOptions.
 * Fixed displaying toast in desktops. Added missing support for desktops in toast handler.
+* Fixed dialog and page report modes not triggering rejection of report when clicking on back button.
 
 ## 0.6.5
 * Changed package_info to package_info_plus to support all platforms.