+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Added filterFunction to CatcherOptions.
 * Fixed displaying toast in desktops. Added missing support for desktops in toast handler.
 * Fixed dialog and page report modes not triggering rejection of report when clicking on back button.
+* Added shouldHandleWhenRejected() method in report handler. Should handle rejected
+* Added handleWhenRejected parameter to ConsoleHandler, ToastHandler, FileHandler.
 
 ## 0.6.5
 * Changed package_info to package_info_plus to support all platforms.